requirement.txt
Django==2.2.4
djangorestframework==3.10.2
mysqlclient==1.4.4
pytz==2019.2
sqlparse==0.3.0
virtualenv==16.7.4
